@@ -20,21 +20,21 @@
     "crx-to-zip.js"
   ],
   "dependencies": {
-    "acorn": "8.2.4",
+    "acorn": "8.8.1",
     "base-64": "1.0.0",
-    "crypto-js": "4.0.0",
-    "fetch-enhanced": "6.0.0",
-    "node-fetch": "2.6.1",
-    "parse5": "6.0.1",
-    "postcss": "8.2.15",
+    "crypto-js": "4.1.1",
+    "fetch-enhanced": "10.1.0",
+    "node-fetch": "3.3.0",
+    "parse5": "7.1.1",
+    "postcss": "8.4.19",
     "rfdc": "1.3.0",
     "unzipper": "0.10.11",
-    "url-toolkit": "2.2.2"
+    "url-toolkit": "2.2.5"
   },
   "devDependencies": {
-    "eslint": "7.26.0",
-    "eslint-config-silverwind": "33.0.0",
-    "updates": "12.1.0",
-    "versions": "8.4.7"
+    "eslint": "8.28.0",
+    "eslint-config-silverwind": "62.0.1",
+    "updates": "13.2.1",
+    "versions": "10.2.4"
   }
 }
